cart 추가, 수정, 삭제 완료
#본문
cart 추가, 수정, 삭제 완료

+++ src/main/java/com/choi/springshop/application/dto/Product/GetProductListRequest.java
@@ -11,6 +11,6 @@ public class GetProductListRequest {
 
     private String imageUrl;
 
-    private Float price;
+    private Integer price;
 
 }
